chore: add audit check to E2E tests
Add tests that validates the audit process if working as expected.

The test deploys a pod before all the Gatekeeper stack that
will trigger violations after the rules are deployed.

The violations should then appear in the `status` field of the constraints.

A second test checks that after the Gatekeeper stack has been deployed
the violations are present.

Fixes #62